Add a selection reader; include selection reader with quickbar in page editor; refactor some context menu code
