+++ src/blocks/readers/SelectionReader.ts
@@ -1,0 +1,73 @@
+/*
+ * Copyright (C) 2022 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/*
+ * Copyright (C) 2022 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { Reader } from "@/types";
+import { ReaderOutput, Schema } from "@/core";
+import selectionController from "@/utils/selectionController";
+
+export class SelectionReader extends Reader {
+  constructor() {
+    super(
+      "@pixiebrix/selection",
+      "Selection Reader",
+      "Data about the current selection"
+    );
+  }
+
+  defaultOutputKey = "selection";
+
+  async isRootAware(): Promise<boolean> {
+    return false;
+  }
+
+  async isAvailable(): Promise<boolean> {
+    return true;
+  }
+
+  outputSchema: Schema = {
+    type: "object",
+    properties: {
+      selectionText: {
+        type: "string",
+        description: "The text for the current selection, if any.",
+      },
+    },
+  };
+
+  async read(): Promise<ReaderOutput> {
+    return {
+      selectionText: selectionController.get(),
+    };
+  }
+}
