+++ src/utils/selectionController.ts
@@ -15,6 +15,33 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+import { getCommonAncestor } from "@/nativeEditor/infer";
+
+/**
+ * Get the HTMLElement corresponding to the current selection.
+ *
+ * Originally introduced to guess which HTML element the user right-clicked to trigger a context menu if PixieBrix
+ * did not already have access to the page.
+ *
+ * @see setActiveElement
+ */
+export function guessSelectedElement(): HTMLElement | null {
+  const selection = document.getSelection();
+  if (selection?.rangeCount) {
+    const start = selection.getRangeAt(0).startContainer.parentNode;
+    const end = selection.getRangeAt(selection.rangeCount - 1).endContainer
+      .parentNode;
+    const node = getCommonAncestor(start, end);
+    if (node instanceof HTMLElement) {
+      return node;
+    }
+
+    return null;
+  }
+
+  return null;
+}
+
 /**
  * Overridable selection getter. Useful to allow the QuickBar to preserve the selection
  * https://github.com/pixiebrix/pixiebrix-extension/issues/2443
